Rearch module service

+++ common/types/module/module.go
@@ -5,7 +5,6 @@ import (
 	"go.uber.org/zap"
 )
 
-// TODO: [v1.x] Retourner une erreur avec type precis (pour la gestion d'erreur)
 type Module interface {
 	Start(queue *NotificationQueue, logger *zap.Logger) error
 	Configure(properties *properties.Properties) error
@@ -16,5 +15,5 @@ type Module interface {
 	StopSession() error
 
 	Name() string
-	State() int
+	State() byte
 }
