+++ common/types/service/service.go
@@ -5,7 +5,6 @@ import (
 	"go.uber.org/zap"
 )
 
-// TODO: [v1.x] Retourner une erreur avec type precis (pour la gestion d'erreur)
 type Service interface {
 	Start(queue *NotificationQueue, logger *zap.Logger) error
 	Configure(properties *properties.Properties) error
@@ -13,5 +12,5 @@ type Service interface {
 	Stop() error
 
 	Name() string
-	State() int
+	State() byte
 }
